Added GenServer to store Feed items/problems.
This is currently just a simple GenServer with two lists. I picked this because ETS tables don't really fit the data we're storing, because it's not a key-value structure.

That said, if performance becomes a problem (copying larger lists between processes), we can escalate to use ETS. Not sure if that will ever happen though.