Timezone can now be configured via ENV variable.
My original idea to get it from the users browser won't work out because there is no standard way to do this, much less without any JavaScript. It must be configured by the user via ENV var then.

When the variable is read, we verify that it's a valid timezone. The rest of the code assumes that it is and crashes if it isn't.