FIX: Atom parser is now smarter when picking from multiple links.
The [Atom Standard](https://www.ietf.org/rfc/rfc4287.txt) allows for multiple links of different types to be added to a single entry. Usually, we want those pointing at the content that the entry is made for. From these, we want the content representation that is viewable in a browser.