FIX: Moved timezone validation from runtime config to Application.
The `Timex` call doesn't work as early as `config/runtime.exs` is evaluated, because the application isn't started yet. In the `Application` though, all dependency apps are started already, so we're free to use it there.

The idea is that an invalid timezone will shutdown the App immediately with an appropriate error message.