Set visible App version to Git release version.
The version displayed in the App is now set (at compiletime) from an `ARG` variable that is passed in from the outside during the Image build.
The Github action reuses the semver label that is assigned to the final image for that.
This in turn is generated from the tagged Github release that started the build.