Set timezone for Cron jobs.
This means that the default/user supplied Timezone is now also used for scheduling the refresh. This gives more expected behavior, especially with daylight savings.

What you usually want is to specify a wall-clock time of when you want new content to be available. Converting this to UTC and adjusting it manually for daylight savings isn't always clear or possible.